Add files via upload
This automated communication system is developed using Python. The project file contains a python script (bot.py). Talking about this chatbot, it allows the user to provide suitable numbers for age guessing and counting. At last, this chatbot asks you multiple-choice questions. Also, this is a simple cmd-based project which is easy to understand and use.